[Move-only] Move active masternode initialization from init.cpp to tiertwo/init.cpp
