+++ src/init.cpp
@@ -90,7 +90,6 @@ static const bool DEFAULT_PROXYRANDOMIZE = true;
 static const bool DEFAULT_REST_ENABLE = false;
 static const bool DEFAULT_DISABLE_SAFEMODE = false;
 static const bool DEFAULT_STOPAFTERBLOCKIMPORT = false;
-static const bool DEFAULT_MASTERNODE  = false;
 static const bool DEFAULT_MNCONFLOCK = true;
 
 std::unique_ptr<CConnman> g_connman;
@@ -1726,46 +1725,8 @@ bool AppInitMain()
     // ********************************************************* Step 10: setup layer 2 data
 
     LoadTierTwo(chain_active_height);
+    if (!InitActiveMN()) return false;
     RegisterTierTwoValidationInterface();
-
-    fMasterNode = gArgs.GetBoolArg("-masternode", DEFAULT_MASTERNODE);
-
-    if ((fMasterNode || masternodeConfig.getCount() > -1) && fTxIndex == false) {
-        return UIError(strprintf(_("Enabling Masternode support requires turning on transaction indexing."
-                                   "Please add %s to your configuration and start with %s"), "txindex=1", "-reindex"));
-    }
-
-    if (fMasterNode) {
-        const std::string& mnoperatorkeyStr = gArgs.GetArg("-mnoperatorprivatekey", "");
-        const bool fDeterministic = !mnoperatorkeyStr.empty();
-        LogPrintf("IS %sMASTERNODE\n", (fDeterministic ? "DETERMINISTIC " : ""));
-
-        if (fDeterministic) {
-            // Check enforcement
-            if (!deterministicMNManager->IsDIP3Enforced()) {
-                const std::string strError = strprintf(_("Cannot start deterministic masternode before enforcement. Remove %s to start as legacy masternode"), "-mnoperatorprivatekey");
-                LogPrintf("-- ERROR: %s\n", strError);
-                return UIError(strError);
-            }
-            // Create and register activeMasternodeManager
-            activeMasternodeManager = new CActiveDeterministicMasternodeManager();
-            auto res = activeMasternodeManager->SetOperatorKey(mnoperatorkeyStr);
-            if (!res) { return UIError(res.getError()); }
-            RegisterValidationInterface(activeMasternodeManager);
-            // Init active masternode
-            const CBlockIndex* pindexTip = WITH_LOCK(cs_main, return chainActive.Tip(); );
-            activeMasternodeManager->Init(pindexTip);
-        } else {
-            // Check enforcement
-            if (deterministicMNManager->LegacyMNObsolete()) {
-                const std::string strError = strprintf(_("Legacy masternode system disabled. Use %s to start as deterministic masternode"), "-mnoperatorprivatekey");
-                LogPrintf("-- ERROR: %s\n", strError);
-                return UIError(strError);
-            }
-            auto res = initMasternode(gArgs.GetArg("-masternodeprivkey", ""), gArgs.GetArg("-masternodeaddr", ""), true);
-            if (!res) { return UIError(res.getError()); }
-        }
-    }
 
     // set the mode of budget voting for this node
     SetBudgetFinMode(gArgs.GetArg("-budgetvotemode", "auto"));
