+++ src/tiertwo/init.h
@@ -5,6 +5,7 @@
 #ifndef PIVX_TIERTWO_INIT_H
 #define PIVX_TIERTWO_INIT_H
 
+static const bool DEFAULT_MASTERNODE  = false;
 
 /** Loads from disk all the tier two related objects */
 bool LoadTierTwo(int chain_active_height);
@@ -16,6 +17,9 @@ void RegisterTierTwoValidationInterface();
 void DumpTierTwo();
 
 void SetBudgetFinMode(const std::string& mode);
+
+/** Initialize the active Masternode manager */
+bool InitActiveMN();
 
 
 #endif //PIVX_TIERTWO_INIT_H
